usb: android: Add HID function driver to android gadget driver
The current Android Gadget Driver does not support HID as one of the
functions.

The HID function driver supports a single device with one boot
interface that does not specify any protocol.

Add HID report descriptor that only states the basic report of
a HID driver and does not support any mouse or keyboard functionalities.

Change-Id: I8c8d73eb63136fd4d49b27e9d0bcf60a15716bd5
Signed-off-by: Kitty Ho <kittyh@codeaurora.org>